+++ static/debug.gd
@@ -22,29 +22,68 @@ extends Object
 
 ## Provides debug static functions.
 ##
-## Print & log functions return true so they can be wrapped in assert(). E.g.,[br]
-##     [code]assert(IVDebug.dlog("debug print"))[/code][br]
-##     [code]assert(!DPRINT or IVDebug.dprint("debug print"))[/code]
+## "dprint", "dlog" and most other functions always return true. They can be
+## wrapped in assert() so they are called only in debug builds:
 ##
-## TODO: Make this a self-contained singleton. Consolidate log_initializer.gd
-## and related IVCoreSettings and IVGlobal members.
+## [codeblock]
+## const DPRINT := true
+##
+## assert(!DPRINT or IVDebug.dprint("print this only if DPRINT == true"))
+## assert(IVDebug.dlog("log this in debug.log"))
+## assert(IVDebug.register_objects_recursive_on_nodes_added())
+## [/codeblock]
+##
+## A debug log is opened only in debug builds. It is flushed on
+## [signal IVGlobal.run_state_changed], which happens whenever the sim stops
+## for a popup (e.g., on [param esc] for the main menu) and closed when project
+## autoloads exit the tree. To prevent log file opening, set
+## [member dlog_name] = ""[br][br]
+##
+## Methods with a large singular output usually create a separate log file
+## specified by [param log_name].
+
 
 static var log_directory := "user://logs"
+static var dlog_name := "debug.log"
 
+static var _dlog: FileAccess
+
+
+# Standard file order is violated below to keep related code together...
+
+# *****************************************************************************
+# log init, flush and destruction
+
+static func _static_init() -> void:
+	if !dlog_name or !OS.is_debug_build():
+		return
+	_dlog = FileAccess.open(log_directory.path_join(dlog_name), FileAccess.WRITE)
+	assert(_dlog, "Failed to open %s" % log_directory.path_join(dlog_name))
+	IVGlobal.run_state_changed.connect(_dlog_flush) # e.g., main menu opened/closed
+	IVGlobal.tree_exited.connect(_dlog_destroy)
+
+
+static func _dlog_flush(_dummy := false) -> void:
+	_dlog.flush()
+
+
+static func _dlog_destroy() -> void:
+	_dlog.close()
+	_dlog = null
+
+
+# *****************************************************************************
+# simple dprint and dlog
 
 static func dprint(arg: Variant, arg2: Variant = "", arg3: Variant = "", arg4: Variant = ""
 		) -> bool:
-	# For >4 items, just use an array.
 	prints(arg, arg2, arg3, arg4)
 	return true
 
 
 static func dlog(arg: Variant) -> bool:
-	var file := IVGlobal.debug_log
-	if !file:
-		return true
-	var line := str(arg)
-	file.store_line(line)
+	if _dlog:
+		_dlog.store_line(str(arg))
 	return true
 
 
@@ -78,14 +117,20 @@ static func dlog_nodes_recursive(node: Node = null, include_internal: bool = tru
 	return true
 
 
+# *****************************************************************************
+# verbose signalling
 
+## Prints when [param signal_name] is emitted from [param object] (print line
+## will be prefixed with [param prefix]). Call before any other connections
+## to this signal so that the print will occur first. Signal must have <= 8 args.
 static func signal_verbosely(object: Object, signal_name: String, prefix: String) -> void:
-	# Call before any other signal connections; signal must have <= 8 args.
 	object.connect(signal_name, IVDebug._on_verbose_signal.bind(prefix + " " + signal_name))
 
 
+## Prints when any signal is emitted from [param object] (print line
+## will be prefixed with [param prefix]). Call before any other connections
+## to this signal so that the print will occur first. Signal must have <= 8 args.
 static func signal_verbosely_all(object: Object, prefix: String) -> void:
-	# See signal_verbosely. Prints all emitted signals from object.
 	var signal_list := object.get_signal_list()
 	for signal_dict in signal_list:
 		var signal_name: String = signal_dict.name
@@ -95,7 +140,8 @@ static func signal_verbosely_all(object: Object, prefix: String) -> void:
 static func _on_verbose_signal(arg: Variant, arg2: Variant = null, arg3: Variant = null,
 		arg4: Variant = null, arg5: Variant = null, arg6: Variant = null, arg7: Variant = null,
 		arg8: Variant = null, arg9: Variant = null) -> void:
-	# Expects signal_name as last bound argument.
+	# Debug <prefix + signal_name> is the last arg (bound after the signal args),
+	# so all nulls after that can be removed.
 	var args := [arg, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9]
 	while args[-1] == null:
 		args.pop_back()
@@ -108,20 +154,30 @@ static func _on_verbose_signal(arg: Variant, arg2: Variant = null, arg3: Variant
 
 static var _object_register: Dictionary[int, WeakRef] = {}
 
+## Registers objects recursively, searching properties (inculding nested in
+## arrays or dictionaries), signal connections, and node children.
+## Objects are kept as weak references so RefCounted will free normally.
+## Supply [param log_name] to log the register now. Call
+## [method log_object_register] to log at a later time. Use to test whether
+## objects that are supposed to free themselves are really doing so.
+static func register_objects_recursive(root: Node, log_name := "") -> bool:
+	_register_object_recursive(root)
+	if log_name:
+		log_object_register(log_name)
+	return true
 
-static func register_objects_on_nodes_added() -> bool:
+
+## Registers objects recursively when any node is added to the tree. The
+## registration is deferred so objects set or connected on _ready() will exist.
+## See comments in [method register_objects_recursive].
+static func register_objects_recursive_on_nodes_added() -> bool:
 	IVGlobal.get_tree().node_added.connect(_on_node_added)
 	return true
 
 
-static func register_all_objects_now(root: Node, log_name := "objects.log") -> bool:
-	_register_object_recursive(root)
-	if log_name:
-		dlog_object_register(log_name)
-	return true
-
-
-static func dlog_object_register(log_name := "objects.log") -> bool:
+## Log the object register now. See [method register_objects_recursive]
+## and [method register_objects_recursive_on_nodes_added].
+static func log_object_register(log_name := "objects.log") -> bool:
 	var file := FileAccess.open(log_directory.path_join(log_name), FileAccess.WRITE)
 	for id in _object_register:
 		var object: Object = _object_register[id].get_ref()